ai/requirements.txt
# AI/ML Dependencies for AgriFinSight

# Core ML frameworks
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# Data processing
numpy==1.24.3
pandas==2.1.3
matplotlib==3.7.2
seaborn==0.12.2

# Model training and evaluation
tensorboard==2.15.1
wandb==0.16.0
mlflow==2.8.1

# Image processing and augmentation
imgaug==0.4.0
scikit-image==0.21.0

# Utilities
tqdm==4.66.1
joblib==1.3.2
pickle5==0.0.12

# Pre-trained models
timm==0.9.12
transformers==4.35.2

# Weather and environmental data
requests==2.31.0
python-dateutil==2.8.2

# Model serving
fastapi==0.104.1
uvicorn==0.24.0
